+++ frontend/src/util/customFetch.js
@@ -1,7 +1,6 @@
-import { getAccessToken, setAccessToken } from '@/auth/accessToken';
-import { getRefreshToken } from '@/auth/refreshToken';
+import { getAccessToken } from '@/auth/accessToken';
 import axios from 'axios';
-import { ENDPOINTS } from './endpoints';
+import { renewAccessToken } from '@/auth/renewAccessToken';
 
 const instance = axios.create({
     baseURL: '/api'
@@ -19,20 +18,6 @@ instance.interceptors.request.use(
   error => Promise.reject(error)
 );
 
-let isRefreshing = false;
-let retryQueue = [];
-
-const processQueue = (error) => {
-  retryQueue.forEach(retry => {
-    if (error) {
-      retry.reject(error);
-    } else {
-      retry.resolve();
-    }
-  });
-  retryQueue = [];
-};
-
 instance.interceptors.response.use(
   response => response,
   async error => {
@@ -40,35 +25,11 @@ instance.interceptors.response.use(
 
     if (error.status === 401 && !originalRequest._retry) {
       originalRequest._retry = true;
-      if (isRefreshing) {
-        return new Promise((resolve, reject) => {
-          retryQueue.push({
-            resolve: () => {
-              resolve(instance(originalRequest));
-            },
-            reject
-          });
-        });
-      }
-
-      isRefreshing = true;
-
       try {
-        const refreshToken = getRefreshToken();
-        if (!refreshToken) {
-          return Promise.reject(error);
-        }
-        const res = await customFetch(ENDPOINTS.auth.refreshToken, {
-          data: { refreshToken }
-        });
-        setAccessToken(res.data.accessToken);
-        processQueue(null);
+        await renewAccessToken();
         return instance(originalRequest);
       } catch (err) {
-        processQueue(err);
-        return Promise.reject(err);
-      } finally {
-        isRefreshing = false;
+        return Promise.reject(error);
       }
     }
